Clean up and added feature
I have cleaned up some of the comments and checked for pep8 styling.

I have also added a funtionality to custom window where we can do the custom patterns either using a click of drag.

By defualt is is by click but to use dragging, hit 'd' on the key board. To stop, hit 's'